fix: split automation files under the max-lines ratchet
Move circuit store methods, host circuit tests, mobile catalog helpers, and detail-sheet actions into sibling files so new automation polish stays under 500 lines.

Amp-Thread-ID: https://ampcode.com/threads/T-01a089af-fecc-72c8-9135-e23981545823
